⚙️Call go() by default on the webpack build
- Do not export the middleware if importing from node
