+++ packages/ink-basic/index.html
@@ -94,7 +94,5 @@ function logToConsole(value, calories) {
         <ink-display bind="v2"></ink-display> m/s<br>
     </ink-aside>
 </ink-scope>
-
-<script>document.addEventListener("DOMContentLoaded", () => ink.go());</script>
 </body>
 </html>
